roles & doctor crud

+++ app/Http/Controllers/MedCategoryController.php
@@ -3,15 +3,30 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Request as HttpRequest;
 
 use App\Models\MedCategory;
 class MedCategoryController extends Controller
 {
     public function index(){
-        $medcategories = MedCategory::orderBy('id')->get();
+        // $medcategories = MedCategory::orderBy('id','desc')->get();
+        // $search = HttpRequest::input('search');
+        // if ($search) {
+        //     $medcategories->where(function ($query) use ($search) {
+        //         $query->where('name', 'like', '%' . $search . '%');
+        //     });
+        // }
+
         return inertia('MedCategory/Index',[
-            'medcategories' => $medcategories,
-        ]);
+            'medcategories' => MedCategory::query()
+                    ->when(HttpRequest::input('search'), function ($query, $search) {
+                        $query->where('name', 'like', '%' . $search . '%');
+                    })->paginate(8)
+                    ->withQueryString(),
+                    'filters' => HttpRequest::only(['search'])
+            ]);
+
+
     }
 
     // public function create(){
@@ -40,6 +55,13 @@ class MedCategoryController extends Controller
     }
 
     public function destroy(MedCategory $medcategory) {
+
+        $medc= $medcategory->medicine()->count();
+
+        if( $medc > 0 ){
+            return back()->withErrors(['GeneralErrors' => "You cannot delete the medicine category $medc->name because it has $medc medicines ."]);
+        }
+
         $medcategory->delete();
 
         return back();
